libutil: remove readInt, readLongLong
the former isn't even int, it's unsigned. the latter is uint64. both
should be explicit about this to avoid problems that we have already
had, such as the FramedSource wire protocol writing using 64 bits to
write frame sizes, but 32 bits to read them. large frames will cause
the reading end to crash with an unnecessary serialization exception

Change-Id: I3c15e911f649eec719d4b1c135dde1b6ba020271