libstore/local-derivation-goal: create reference maps between outputs outside of topoSort()
That way we do now have linear complexity to determine output
references per output within each step of the topological sort.

Instead, this is done before and the topo-sort only filters the output
map for other derivation outputs.

Following up on this, we can re-use `outputGraph` to generate a tree
with references to display which files cause an output reference cycle
if needed.

Change-Id: Ibdd46e7b2e895bfeeebc173046d1297b41998181